added google task settings in onefm general settings
