+++ one_fm/one_fm/doctype/onefm_general_setting/onefm_general_setting.js
@@ -6,6 +6,8 @@ frappe.ui.form.on('ONEFM General Setting', {
 		if(!frm.is_new()){
 			frm.trigger('setup_face_recognition');
 		}
+		frm.fields_dict['sync_google_tasks_btn'].df.hidden = !frm.doc.google_task_synchronization_enabled;
+        frm.refresh_field('sync_google_tasks_btn');
 	},
 	setup_face_recognition(frm){
 		frm.add_custom_button('Setup Face Recognition', ()=>{
@@ -15,6 +17,14 @@ frappe.ui.form.on('ONEFM General Setting', {
 				frappe.msgprint(res.message.message)
 			})
 		}, 'Actions')
-		
-	}
+	},
+	sync_google_tasks_btn: function (frm) {
+        if(frm.doc.google_task_synchronization_enabled) {
+            frappe.call({
+                method: 'one_fm.overrides.todo.sync_google_tasks_with_todos'
+            }).then(res => {
+                frappe.msgprint(res.message.message)
+            })
+        }
+    }
 });
