--- one_fm/one_fm/doctype/onefm_google_settings/onefm_google_settings.py
@@ -1,9 +1,0 @@
-# Copyright (c) 2025, omar jaber and contributors
-# For license information, please see license.txt
-
-# import frappe
-from frappe.model.document import Document
-
-
-class ONEFMGoogleSettings(Document):
-	pass
