--- one_fm/one_fm/doctype/onefm_google_settings/test_onefm_google_settings.py
@@ -1,9 +1,0 @@
-# Copyright (c) 2025, omar jaber and Contributors
-# See license.txt
-
-# import frappe
-from frappe.tests.utils import FrappeTestCase
-
-
-class TestONEFMGoogleSettings(FrappeTestCase):
-	pass
